feat: export react context, return wallet session on wallet connect and break circular deps
