+++ packages/client/src/client/actions.ts
@@ -19,16 +19,10 @@ import { fetchNonce } from '@solana-program/system';
 
 import { createLogger, formatError } from '../logging/logger';
 import { createSolanaRpcClient } from '../rpc/createSolanaRpcClient';
-import type {
-	AddressLookupTableData,
-	ClientActions,
-	ClientState,
-	ClientStore,
-	NonceAccountData,
-	SolanaClientRuntime,
-	WalletRegistry,
-} from '../types';
+import type { SolanaClientRuntime } from '../rpc/types';
+import type { AddressLookupTableData, ClientActions, ClientState, ClientStore, NonceAccountData } from '../types';
 import { now } from '../utils';
+import type { WalletRegistry, WalletSession } from '../wallet/types';
 
 type MutableRuntime = SolanaClientRuntime;
 
@@ -162,12 +156,12 @@ export function createActions({ connectors, logger: inputLogger, runtime, store 
 	 * Initiates a wallet connection using a registered connector.
 	 *
 	 * @param connectorId - Identifier for the desired wallet connector.
-	 * @returns Promise that resolves once the connection attempt has completed.
+	 * @returns Promise that resolves to the wallet session once the connection attempt has completed.
 	 */
 	async function connectWallet(
 		connectorId: string,
 		options: Readonly<{ autoConnect?: boolean }> = {},
-	): Promise<void> {
+	): Promise<WalletSession> {
 		walletEventsCleanup?.();
 		walletEventsCleanup = undefined;
 		const connector = connectors.get(connectorId);
@@ -206,6 +200,7 @@ export function createActions({ connectors, logger: inputLogger, runtime, store 
 				level: 'info',
 				message: 'wallet connected',
 			});
+			return session;
 		} catch (error) {
 			store.setState((state) => ({
 				...state,
