+++ packages/client/src/features/spl.ts
@@ -31,8 +31,9 @@ import {
 } from '@solana-program/token';
 
 import { createTokenAmount, type TokenAmountMath } from '../numeric/amounts';
+import type { SolanaClientRuntime } from '../rpc/types';
 import { createWalletTransactionSigner, isWalletSession, resolveSignerMode } from '../signers/walletTransactionSigner';
-import type { SolanaClientRuntime, WalletSession } from '../types';
+import type { WalletSession } from '../wallet/types';
 import type { SolTransferSendOptions } from './sol';
 
 type BlockhashLifetime = Readonly<{
