+++ packages/client/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"module": "ESNext"
+	},
+	"include": ["src/**/*.test.ts", "src/**/*.test.tsx"],
+	"exclude": []
+}
