+++ packages/react-hooks/package.json
@@ -34,7 +34,8 @@
 		"lint": "biome check src",
 		"test:typecheck": "tsc --noEmit",
 		"test": "vitest run --config ../../vitest.config.ts",
-		"typecheck": "pnpm test:typecheck"
+		"typecheck": "pnpm test:typecheck",
+		"check:circular-deps": "madge --circular $(find ./src -name '*.ts' -o -name '*.tsx')"
 	},
 	"author": "Solana Maintainers <maintainers@solana.foundation>",
 	"repository": {
