+++ packages/react-hooks/src/hooks.ts
@@ -237,7 +237,7 @@ export function useWalletActions() {
 export function useConnectWallet(): (
 	connectorId: string,
 	options?: Readonly<{ autoConnect?: boolean; allowInteractiveFallback?: boolean }>,
-) => Promise<void> {
+) => Promise<WalletSession> {
 	const client = useSolanaClient();
 	return useCallback(
 		(connectorId: string, options?: Readonly<{ autoConnect?: boolean; allowInteractiveFallback?: boolean }>) =>
