+++ turbo.json
@@ -22,6 +22,9 @@
 		"typecheck": {
 			"dependsOn": ["^build"],
 			"outputs": []
+		},
+		"check:circular-deps": {
+			"outputs": []
 		}
 	}
 }
